short-url-application/application/auth/app_files/requirements.txt
Flask
Flask-MySQLdb
mysql-connector-python
PyMySQL
cryptography
prometheus_client
opentelemetry-distro
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-pymysql
prometheus-flask-exporter==0.23.2